render: fix glitches at high zoom levels (e.g. 19+)
Using the benchmark suite introduced recently, we can see that this fix
has an impact on the performance (as expected, boolean ops aren't free)
    group                    baseline          fixed
    -----                    --------          -----
    Render/Empty/Carved        128.6±2.28ns      131.1±3.35ns        1.02
    Render/Empty/Normal         85.4±2.22ns       87.7±4.25ns        1.03
    Render/Few/Carved           54.0±0.33µs       98.6±0.80µs        1.82
    Render/Few/Normal           50.3±0.40µs       90.7±0.60µs        1.80
    Render/Full/Carved          17.4±0.19µs       26.8±0.23µs        1.54
    Render/Full/Normal          16.8±0.16µs       25.6±0.18µs        1.52
    Render/Many/Carved       1956.1±18.26µs        3.3±0.02ms        1.69
    Render/Many/Normal       1633.1±21.83µs        3.0±0.04ms        1.82
    Render/Sparse/Carved       630.6±4.07µs    1071.8±16.72µs        1.70
    Render/Sparse/Normal       515.8±7.81µs     937.2±23.84µs        1.82

So we end up with 50 to 80% of overhead, which seems crazy high but when
looking at the actual values the timings are still good I believe.
=> sub-ms in most cases, except for tiles with lot of cell indexes (we
are at 3 ms but those were already 2ms to begin with).

Moreover, the fix also decrease the size of the generated tiles by 25%
on average, so we may get back some of the lost time at decoding and
rendering time.

And if it ever becomes an issue in practice, we can still make this pass
opt-in and let the caller enable it as needed (e.g. disabled for low
zoom and enabled for high zoom).

But let's keep things simple for now.