bump h3o to 0.7
Had to temporarily disable a unit test.

Not 100% sure it's a geozero bug (https://github.com/georust/geozero/issues/218),
it could also be on our side, but the generated MVT does look visually
good and can be decoded by other libraries, so it seems to be working
fine on our side.